fastapi/requirements.txt
fastapi
uvicorn
langchain-openai
langchain-core
langsmith
python-dotenv
pypdf
pydantic
